fix error when running on Windows with no ctypes module
The problem has been encountered using 64-bit CPython 2.5.4 that does not
include the ctypes module by default.

There AnsiToWin32 object's win32_calls attribute would get initialized to None
if ctypes is not available. However, later in AnsiToWin32.call_win32() this
attribute gets iterated over as if it were a dict. Fix is to initialize it to
an empty dictionary in case Windows APIs can not get called.