Guard against multiple calls to init().
BEWARE, I haven't tested this on Windows (because I
don't have a Windows box.) It should definitely get
tested there before we merge.

This should no longer lose the original stdout/stderr,
so that we can restore them with a single call to deinit(),
while still preserving the behavior that multiple calls
to init() can override the passed values for autoreset,
convert, strip, wrap.

This fix is not intended to reset output to default
foreground/background colors after the call to
'deinit()' -  that was never intended behavior.