Move package to a top-level "src" directory
This helps ensure that tox tests the colorama installed to the
virtualenv and not otherwise working due to the current working
directory.

The tests directory was moved out of the source directory to the top
level. Test files don't need to be included in the installed package.
They exists for developers, not end users. They still ship with the
sdist so downstream packages can run the tests when re-packaging
Colorama.

The demos directory has been updated to work with the new directory
structure. Its use is now documented in README.rst.

Switch test runner to pytest in tox.ini. pytest has been moved out of
the lint_python GitHub workflow as it is now covered by the test
workflow.

Fixes #278